src/requirements.txt
pytest
playwright
pytest-playwright
pytest-xdist
allure-pytest
PyYAML